fix(audio): keep media-session rate conversion out of the audio ISR
Bound media (AX) sessions at non-48 kHz ran the 44.1->48 kHz polyphase
FIR inside audio_fabric_isr. On core 0 that ISR shares the GIC priority
and core with isr_video, so every collision delayed the vblank handler
past the overlay rearm window: the per-vblank overlay VDMA reset landed
in active video, bumping the generation and invalidating both line
buffer banks, which showed as windowed-PIP flicker (zz9000-drivers#83,
reproducible with ZZPlay and any MPEG-1 video whose audio binds to AX;
--audio=ahi and --audio=none were clean, and RC2 with the linear
resampler was clean).

Media sessions now stage converted 48 kHz PCM through the existing
preconvert ring on the main loop, exactly like bound audio streams, so
the ISR fill only copies. The exact view<->session cursor mapping lives
in audio_pump_media_view.h: whole 3840-byte view periods map to whole
source periods (882 frames / 3528 bytes at 44.1 kHz stereo), retirement
is clamped to what the converter really consumed so the zero-padded EOS
tail retires exactly, and the preconvert ring owner is tagged by kind
because media and stream session ids are independent namespaces.

The preconvert source cursors widen to uint64_t so media session
cursors (already 64-bit) map without truncation.

overlay.c gains only a comment: the per-vblank rearm reset is the
overlay stream's phase lock to the raster and must stay unconditional
(a gating experiment confirmed drift artifacts when it is skipped).

Hardware-verified: flicker gone with ZZPlay + AX audio at 1280x1024/32.

+++ ZZ9000_proto.sdk/ZZ9000OS/src/audio_pump_preconvert.h
@@ -21,8 +21,8 @@
 struct audio_pump_preconvert_source {
 	uint8_t *ring;
 	uint32_t capacity;
-	uint32_t produced;
-	uint32_t consumed;
+	uint64_t produced;
+	uint64_t consumed;
 	uint32_t sample_rate;
 	uint32_t channels;
 	uint32_t sample_format;
@@ -46,7 +46,7 @@ uint32_t audio_pump_preconvert_used(
 	const struct audio_pump_preconvert *state);
 int audio_pump_preconvert_fill(struct audio_pump_preconvert *state,
 	const struct audio_pump_preconvert_source *source,
-	uint32_t *source_consumed);
+	uint64_t *source_consumed);
 int audio_pump_preconvert_stage(struct audio_pump_preconvert *state,
 	uint32_t bytes);
